requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.9.0
altair==5.3.0
asttokens==2.4.1
attrs==23.2.0
av==12.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.1
executing==2.0.1
facenet-pytorch==2.6.0
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
grpcio==1.64.1
idna==3.7
ifaddr==0.2.0
ipykernel==6.29.4
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
online-triplet-loss==0.0.6
opencv-python==4.8.1.78
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pymongo==4.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.36.0
sympy==1.12.1
tenacity==8.4.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torch-tb-profiler==0.4.3
torcheval==0.0.7
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
